Fixed service_type => 16bits in gxscomments. Removed OLD RelatedMessage() data checks - which we killing comment tree. Added debug indicating comment / vote counts.
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@6191 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ libretroshare/src/services/p3gxscommon.h
@@ -37,7 +37,7 @@ class p3GxsCommentService
 {
 	public:
 
-	p3GxsCommentService(RsGenExchange *exchange, uint32_t service_type);
+	p3GxsCommentService(RsGenExchange *exchange, uint16_t service_type);
 
 	bool getGxsCommentData(const uint32_t &token, std::vector<RsGxsComment> &msgs);
 	bool getGxsRelatedComments(const uint32_t &token, std::vector<RsGxsComment> &msgs);
@@ -52,7 +52,7 @@ class p3GxsCommentService
 	private:
 
 	RsGenExchange 	*mExchange;
-	uint32_t	mServiceType;
+	uint16_t	mServiceType;
 };
